Refactoring API calls to server dir
