+++ server/routes/relatedItems_routes.js
@@ -13,6 +13,81 @@ router.use(fileUpload());
 router.use(bodyParser.json());
 router.use(bodyParser.urlencoded({extended: true}));
 
+const product_id = 22134;
+
+router.get('/relatedItems', (req, res) => {
+  // Turn this line on when app goes live
+  // const builtUrl = url + `/products/${req.data.product_id}/related`;
+
+  // Turn this line off when app goes live
+  const builtUrl = url + `/products/22134/related`;
+
+  const options = {
+    method: 'get',
+    url: builtUrl,
+    headers: token.AUTH
+  };
+
+  axios(options)
+    .then(({ data }) => {
+      return Promise.all(data.map((id) => {
+
+        const builtUrl = url + `/products/${id}`;
+
+        const options = {
+          method: 'get',
+          url: builtUrl,
+          headers: token.AUTH
+        }
+
+        return axios(options)
+          .then(({ data }) => {
+            itemData[data.id] = data;
+          });
+      }));
+    }).then((itemData) => {
+      res.send(itemData);
+    }).catch((err) => {
+      console.log(err);
+    });
+});
+
+// const getRelatedItemsIds = (product_id) => {
+//   const builtUrl = url + `/products/${product_id}/related`;
+
+//   const options = {
+//     method: 'get',
+//     url: builtUrl,
+//     headers: token.AUTH
+//   };
+
+//   axios(options)
+//     .then(({ data }) => {
+//       return Promise.all(data.map((id) => {
+
+//         return this.getRelatedItemsData(id);
+
+//       }));
+//     }).then((data) => {
+//       res.send(data);
+//     }).catch((err) => {
+//       console.log(err);
+//     });
+// };
+
+// const getRelatedItemsData = (id) => {
+//   const builtUrl = url + `/products/${id}`;
+//   const options = {
+//     method: 'get',
+//     url: builtUrl,
+//     headers: token.AUTH
+//   }
+//   return axios(options)
+//     .then(({ data }) => {
+//       this.data[data.id] = data;
+//     });
+// };
+
 // // NON-API Photo upload Logic (TomHo)
 // router.post('/uploadphoto', (req, res) => {
 //   console.log('upload post received');
